Изменил форму адреса

+++ src/app/admin/dashboard/address/form/address-form.component.ts
@@ -2,9 +2,10 @@ import { Component, Inject, OnInit } from "@angular/core";
 import { NgForm } from "@angular/forms";
 import { MatDialogRef, MAT_DIALOG_DATA } from "@angular/material/dialog";
 import { AddressService } from "src/app/shared/services/address.service";
-import { Address } from "src/app/shared/interface";
+import { Address, Place } from "src/app/shared/interface";
 import { EventsForm } from "src/app/admin/shared/interface";
 import { isEmptyObject } from "src/app/admin/shared/utils";
+import { ConfirmDialogService } from "src/app/admin/shared/components/confirm/confirm.service";
 
 @Component({
     selector: 'dialog-address',
@@ -27,14 +28,19 @@ export class DialogAddressComponent implements OnInit, EventsForm {
         longitude: ''
     };
 
+    places: Place[] = [];
+
     constructor(
         public dialogRef: MatDialogRef<DialogAddressComponent>,
         @Inject(MAT_DIALOG_DATA) public data: Address,
-        private serviceAddress: AddressService
+        private serviceAddress: AddressService,
+        private confirmDialog: ConfirmDialogService
     ) { }
 
     ngOnInit(): void {
         this.address = Object.assign(this.address, this.data);
+        this.places = this.data && this.data.places || [];
+
         if (!isEmptyObject(this.data)) {
             this.isReading = true;
             this.isDisabledField = true;
@@ -42,15 +48,15 @@ export class DialogAddressComponent implements OnInit, EventsForm {
     }
     submit(formAddress: NgForm) {
         if (this.isReading) {
-            this.serviceAddress.updateAddress({id: this.address.id, ...formAddress.value}).subscribe(
-                (address) =>  this.dialogRef.close(address?.message),
-                (err) =>  this.dialogRef.close(err)
+            this.serviceAddress.updateAddress({ id: this.address.id, ...formAddress.value }).subscribe(
+                (address) => this.dialogRef.close(address?.message),
+                (err) => this.dialogRef.close(err)
             );
         } else {
             this.serviceAddress.createAddress(formAddress.value)
                 .subscribe(
                     (address) => this.dialogRef.close(address),
-                    (err) =>  this.dialogRef.close(err)
+                    (err) => this.dialogRef.close(err)
                 );
         }
     }
@@ -60,9 +66,15 @@ export class DialogAddressComponent implements OnInit, EventsForm {
     }
 
     deleting(): void {
-        this.serviceAddress.deleteAddress(this.address.id).subscribe(
-            (address) => this.dialogRef.close(address?.message),
-            (err) => this.dialogRef.close(err)
+        this.confirmDialog.confirm('Хотите удалить запись?', 'Да', 'Нет').afterClosed().subscribe(
+            result => {
+                if (result) {
+                    this.serviceAddress.deleteAddress(this.address.id).subscribe(
+                        (address) => this.dialogRef.close(address?.message),
+                        (err) => this.dialogRef.close(err)
+                    );
+                }
+            }
         );
     }
